+++ spec/fixtures/minions-lib/Classes/Minions.h
@@ -1,0 +1,6 @@
+
+#import <Foundation/Foundation.h>
+
+/** Minions are awesome */
+@interface MinionsObj : NSObject
+@end